combining shortwave radiation solar and clouds into one file
