+++ jcm/shortwave_radiation.py
@@ -97,3 +97,55 @@ def clouds(qa ,rh,precnv,precls,iptop,gse,fmask):
     clstr = clstr + fmask * (clstrl - clstr)
 
     return icltop, cloudc, clstr, qcloud
+
+def solar(tyear):
+    """
+    Calculate the daily-average insolation at the top of the atmosphere as a function of latitude.
+    
+    Parameters:
+    tyear : float
+        Time as a fraction of the year (0-1, where 0 corresponds to January 1st at midnight).
+
+    Returns:
+    topsr : array-like
+        Daily-average insolation at the top of the atmosphere for each latitude band.
+    """
+    from jcm.geometry import coa, sia
+    csol = 1368.0
+    
+    # Constants and precomputed values
+    pigr = 2.0 * jnp.arcsin(1.0)
+    alpha = 2.0 * pigr * tyear
+    
+    # Calculate declination angle and Earth-Sun distance factor
+    ca1 = jnp.cos(alpha)
+    sa1 = jnp.sin(alpha)
+    ca2 = ca1**2 - sa1**2
+    sa2 = 2.0 * sa1 * ca1
+    ca3 = ca1 * ca2 - sa1 * sa2
+    sa3 = sa1 * ca2 + sa2 * ca1
+
+    decl = (0.006918 - 0.399912 * ca1 + 0.070257 * sa1 - 
+            0.006758 * ca2 + 0.000907 * sa2 - 
+            0.002697 * ca3 + 0.001480 * sa3)
+
+    fdis = 1.000110 + 0.034221 * ca1 + 0.001280 * sa1 + 0.000719 * ca2 + 0.000077 * sa2
+
+    cdecl = jnp.cos(decl)
+    sdecl = jnp.sin(decl)
+    tdecl = sdecl / cdecl
+
+    # Compute daily-average insolation at the top of the atmosphere
+    csolp = csol / pigr
+
+    # Calculate the solar radiation at the top of the atmosphere for each latitude
+    ch0 = jnp.clip(-tdecl * sia / coa, -1.0, 1.0)
+    h0 = jnp.arccos(ch0)
+    sh0 = jnp.sin(h0)
+
+    topsr = csolp * fdis * (h0 * sia * sdecl + sh0 * coa * cdecl)
+    
+
+    return topsr
+
+
